chore: move proposals components into correct dir
