+++ apps/explorer/src/styles.css
@@ -32,9 +32,9 @@
   line-height: calc(min(var(--ag-line-height, 26px), 26px) - 4px);
 }
 
-.vega-ag-grid .ag-row {
-  border-width: 1px 0;
-  border-bottom: 1px solid transparent;
+.vega-ag-grid .ag-row,
+.vega-ag-grid .ag-cell {
+  border-width: 0;
 }
 
 /* Light variables */
