+++ libs/proposals/src/components/index.ts
@@ -2,3 +2,5 @@ export * from './asset-proposal-notification';
 export * from './market-proposal-notification';
 export * from './protocol-upgrade-countdown';
 export * from './protocol-upgrade-proposal-notification';
+export * from './proposals-list';
+export * from './voting-progress';
