+++ libs/proposals/src/components/proposals-list/index.ts
