+++ libs/proposals/src/components/voting-progress/index.ts
