+++ libs/proposals/src/components/voting-progress/vote-progress.spec.tsx
