+++ libs/proposals/src/components/voting-progress/voting-progress.tsx
