+++ libs/proposals/src/lib/index.ts
@@ -1,6 +1,4 @@
-export * from './proposals-hooks';
-export * from './voting-hooks';
 export * from './proposals-data-provider';
-export * from './proposals-list';
-export * from './voting-progress';
+export * from './proposals-hooks';
 export * from './protocol-upgrade-proposals';
+export * from './voting-hooks';
