+++ libs/react-helpers/src/hooks/index.ts
@@ -13,3 +13,4 @@ export * from './use-yesterday';
 export * from './use-previous';
 export * from './use-logger';
 export * from './use-pane-layout';
+export * from './use-copy-timeout';
